move implementation of `Ligero` to data_structures
